# Make `getindex` for `Bidiagonal` O(1) in the worst-case